ES-5574 Replaced Long with Int for minimumRenewals and maximumRenewals

+++ src/main/kotlin/com/procurement/orchestrator/domain/model/lot/Renewal.kt
@@ -22,10 +22,10 @@ data class Renewal(
     @param:JsonProperty("description") @field:JsonProperty("description") val description: String?,
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    @param:JsonProperty("minimumRenewals") @field:JsonProperty("minimumRenewals") val minimumRenewals: Long?,
+    @param:JsonProperty("minimumRenewals") @field:JsonProperty("minimumRenewals") val minimumRenewals: Int?,
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    @param:JsonProperty("maximumRenewals") @field:JsonProperty("maximumRenewals") val maximumRenewals: Long?,
+    @param:JsonProperty("maximumRenewals") @field:JsonProperty("maximumRenewals") val maximumRenewals: Int?,
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @param:JsonProperty("period") @field:JsonProperty("period") val period: Period?
